common: fix trivial warnings in external headers of PhysX3 and gfx4
replacing: asin->asinf, acos->acosf, atan->atanf, atan2->atan2f,
tan->tanf, log->logf, fabs->fabsf, sqrt->sqrtf

when the result is needed as a float instead of double

+++ dev/external/gfx4/Src/Render/Render_HAL.h
@@ -625,7 +625,7 @@ public:
     { 
         if (sParams.DisplayWidthCm == 0)
         {
-            sParams.DisplayWidthCm = sParams.DisplayDiagInches / sqrt(1.0f + 1.f/sParams.DisplayAspectRatio * 
+            sParams.DisplayWidthCm = sParams.DisplayDiagInches / sqrtf(1.0f + 1.f/sParams.DisplayAspectRatio *
             1.f/sParams.DisplayAspectRatio) * 2.54f /* inches to cm */; 
         }
         if (Matrices->S3DImpl)
